pyproject.toml
[tool.poetry]
name = "feast"
version = "1.0.0"
description = "build feature store from scratch"
authors = ["contact.nampham.work <contact.nampham.work@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "3.9.12"
selenium = "^4.21.0"
beautifulsoup4 = "^4.12.3"
pandas = "^2.2.2"
loguru = "^0.7.2"
minio = "^7.2.7"
pyyaml = "^6.0.1"
deltalake = "^0.18.0"
numpy = "^1.26.4"
psycopg2-binary = "^2.9.9"
sqlalchemy = "^2.0.30"
confluent-kafka = "^2.4.0"
python-dotenv = "^1.0.1"
catboost = "^1.2.5"
fastapi = "^0.111.0"
joblib = "^1.4.2"
uvicorn = {extras = ["standard"], version = "^0.30.1"}
great-expectations = "^0.18.15"
pydeequ = "^1.3.0"
kafka-python = "^2.0.2"
pyspark = "^3.5.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
